extract poki serialization functionality from pali
this commit factors the poki serialization functionality out of pali, and moves it into the poki library. a future commit will be responsible for actually removing the poki functionality from pali, and reimplementing the final assembly and serialization stages using the new library code. a future commit will also implement the deserialize method, making it possible to deserialize poki files as well