introduce separate thread for lua evaluation
this is a fairly rough (i.e. i find the ui-related code quite ugly) but functional implementation of a fairly important change: this commit moves lua evaluation, which was previously done in the same thread as ui rendering and input processing, to its own thread. this change allows for the ui to be continuously rendered and for input to be continuously parsed even while some chunk of lua is actively being evaluated